ENABLED try_subscribe in dsl_communication (read locks were not acquired before)
changed theTime() for wtime()

+++ bmarks/microbench/linkedlist/linkedlist.h
@@ -50,7 +50,8 @@
 
 #define TRANSACTIONAL                   d->unit_tx
 
-typedef intptr_t val_t;
+//typedef intptr_t val_t;
+typedef unsigned long long int val_t;
 typedef unsigned long long int nxt_t;
 
 #define VAL_MIN                         INT_MIN
